Update __init__.py
Added a default-on option to choose what sensors to monitor, based on which sensors are loaded when the print is started.  This bypasses the ability to cancel prints if no filament is loaded, but also prevents T0-only prints with filament not loaded in T1 from canceling the print.

Also added a first pass at checking the filament state (again based on what was present when the print started) on print resume, and pausing again if filament isn't loaded.